Properly load map visualizations from API
Still WIP:
- loading and displaying the map data
- bug with map selection state with years and data sources not being selected properly

+++ backend/src/model/data.rs
@@ -6,7 +6,7 @@ use sqlx::FromRow;
 pub struct Data {
     pub county_id: i16,
     pub state_id: i16,
-    pub source: i16,
+    pub source: i32,
     pub start_date: NaiveDate,
     pub end_date: NaiveDate,
     pub value: Option<f64>,
@@ -14,7 +14,7 @@ pub struct Data {
 
 #[derive(FromRow, Deserialize, Serialize)]
 pub struct SourceAndDate {
-    pub source: i16,
+    pub source: i32,
     pub start_date: NaiveDate,
     pub end_date: NaiveDate,
 }
